requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
langdetect==1.0.9
supabase==2.3.1
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
requests==2.31.0
aiohttp==3.9.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
pytest==8.0.0
httpx==0.26.0
